requirements.txt
google-generativeai>=0.8.0
edge-tts>=6.1.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
pydub>=0.25.1
Pillow>=10.0.0
